Read multiple PUFP .txt files

ufp_batch_read() imports and cleans multiple PUFP text files as a single data frame. Measurements from different monitoring sessions are denoted by the 'Sampling_Event' column, which is defined by event_threshold.

ufp_batch_read(
  paths,
  event_threshold = 10,
  tz = "America/New_York",
  truncate_ufp = TRUE,
  coords = TRUE
)

Arguments

paths
a vector of paths

event_threshold
numeric; difftime value in minutes to differentiate sampling events. Consecutive measurements that exceed this threshold are separated into new sampling events. Default = 10.

tz
a character string that specifies which time zone to parse the date with. Default = 'America/New_York.'

truncate_ufp
logical; truncate UFP concentration? If TRUE (the default), UFP concentrations above 250K will be right censored.

coords
logical; arse GPS string to derive latitude and longitude? Default = TRUE.

Value

a tibble. Additional columns are created for 'Sampling_Event' and 'Sampling_Day.' Sampling events represent discreet intervals in sampling defined by the event_threshold. Sampling day is derived from the date of the measurements. Multiple sampling events may occur within a sampling day.

Examples

if (FALSE) {

ufp_batch_read(c(path1, path2, path3),
  event_threshold = 10,
  tz = "America/New_York", truncate_ufp = TRUE, coords = TRUE
)
}